app/requirements.txt
flask
gunicorn
pq
psycopg2
Flask-SQLAlchemy
git+https://github.com/sshh12/llm_oracle
